✨ feat(suspense & errorbound): suspense & errorbound
applied suspense & errorBound pattern to get rid of verbouse loading and error handling
