+++ src/pages/Main.tsx
@@ -1,94 +1,11 @@
-import React from 'react';
-import WeatherMain from 'components/WeatherMain';
-import styled from 'styled-components';
+import React, { Suspense } from 'react';
 import { Link } from 'react-router-dom';
-import { hourlyWeatherType} from 'types/apiTypes';
 import Stack from 'components/Stack';
 import Pill from 'components/Pill';
-import { useQuery } from 'react-query';
-import { getHourlyWeather, getCurrentWeather, getLocation, getCurrentArea} from 'api/getWeatherData';;
-
-
-const WeatherScore = styled.div`
-  margin-top: 15px;
-  background-color: #fff;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 70px;
-  margin-bottom:20px;
-  border-radius: 15px;
-  font-family: AppleSDGothicNeoB00;
-  font-size: 13pt;
-  & strong {
-    font-size: 20pt;
-    color: #1814af;
-  }
-`;
-
-const WeatherScoreListContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 100px;
-  margin-top:15px;
-  flex-wrap: nowrap;
-  overflow-x: auto;
-  font-family: AppleSDGothicNeoB00;
-  scrollbar-width: none;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  & > .mini-calendar {
-    display: inline-flex;
-    flex-direction: column;
-    align-items: center;
-    width: 65px;
-    flex-shrink: 0;
-    border-radius: 14px;
-    height: 87px;
-    &:not(:last-child) {
-      margin-right: 10px;
-    }
-    & > div {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-    & > .date {
-      background-color: #dbdbdb;
-      height: 34px;
-      width: 100%;
-      font-size: 16px;
-      border-top-left-radius: 14px;
-      border-top-right-radius: 14px;
-    }
-    & > .score {
-      color: #001f8e;
-      height: 40px;
-      background-color: #fff;
-      width: 100%;
-    }
-    & > .temp {
-      color: #000000;
-      height: 30px;
-      background-color: #fff;
-      width: 100%;
-      border-bottom-left-radius: 14px;
-      border-bottom-right-radius: 14px;
-    }
-  }
-`;
+import { WeatherScoreList } from 'components/WeatherScoreList';
+import { ErrorBoundary } from 'react-error-boundary';
 
 const Main = () => {
-  const pos = getLocation();
-  const {isLoading : addressIsLoading, error:addressError, data:addressData} = useQuery(['addressData',pos],()=>getCurrentArea(pos));
-  let area = "대구 북구" // 위치 못찾았을때
-  if(!addressIsLoading && !addressError && typeof addressData !== 'undefined'){
-    area = addressData.address;
-  }
-  const {isLoading:hourlyIsLoading,error:hourlyError,data:hourlyData} = useQuery(["hourlyData",area],()=>getHourlyWeather(area),{enabled:!!addressData});
-  const {isLoading:currentIsLoading,error:currentError,data:currentData} = useQuery(["currentData", area],()=>getCurrentWeather(area),{enabled:!!addressData});
   return (
     <Stack
       style={{
@@ -96,7 +13,7 @@ const Main = () => {
         height: 844,
         padding: 30,
         backgroundColor: '#f5f5f5',
-        borderRadius:30,
+        borderRadius: 30,
         fontFamily: 'AppleSDGothicNeoB00',
       }}
     >
@@ -118,43 +35,17 @@ const Main = () => {
               fontFamily: 'AppleSDGothicNeoB00',
             }}
           >
-            <span style={{fontSize:16,marginLeft:10}}>약속을 정해보세요</span>
+            <span style={{ fontSize: 16, marginLeft: 10 }}>
+              약속을 정해보세요
+            </span>
             <span style={{ position: 'absolute', right: '25px' }}>&gt;</span>
           </Pill>
         </Link>
-        {currentError && <span>error</span>}
-        {currentIsLoading ? <span>loading...</span>
-         : 
-        <WeatherMain
-        locationName={currentData ? currentData.location : ""}
-        weatherCode={currentData ? currentData.weather_main : ""}
-        temperature={currentData ? currentData.current_temp : 0}
-        criteriaTime={currentData ? currentData.current_dt: ""}
-      />
-        }
-        {currentIsLoading ? <span>loading...</span>
-         : 
-        <WeatherScore>
-          오늘의 날씨 점수는 <strong style={{fontSize:20,marginLeft:5}}>{currentData?.todayScore}</strong>점 입니다
-        </WeatherScore>
-        }
-        <span 
-          style={{
-              fontSize:20,
-              fontFamily:'AppleSDGothicNeoB00',
-            }}
-            >시간대별 날씨</span>
-        <WeatherScoreListContainer>
-          {hourlyError && <span>error</span>}
-          {hourlyIsLoading ? <span>loading...</span>
-          : hourlyData?.map((data:hourlyWeatherType, idx:number) => (
-            <div key={idx} className="mini-calendar">
-              <div className="date">{data.dt}</div>
-              <div className="score" style = {{fontSize:20}}>{data.weather}</div>
-              <div className="temp" style = {{fontSize:14}}>{parseInt(data.temp.toFixed())}°C</div>
-            </div>
-          ))}
-        </WeatherScoreListContainer>
+        <ErrorBoundary fallback={<span>Error</span>}>
+          <Suspense fallback={<span>loading</span>}>
+            <WeatherScoreList />
+          </Suspense>
+        </ErrorBoundary>
       </Stack>
     </Stack>
   );
